+++ python/setup.py
@@ -212,7 +212,7 @@ sources = ['pyngraph/function.cpp',
            'pyngraph/ops/softmax.cpp',
            'pyngraph/runtime/backend.cpp',
            'pyngraph/runtime/regmodule_pyngraph_runtime.cpp',
-           'pyngraph/runtime/tensor_view.cpp',
+           'pyngraph/runtime/tensor.cpp',
            'pyngraph/passes/manager.cpp',
            'pyngraph/passes/regmodule_pyngraph_passes.cpp',
            'pyngraph/types/element_type.cpp',
