+++ src/ngraph/runtime/cpu/builder/function_call.cpp
@@ -17,7 +17,7 @@
 #include "ngraph/op/function_call.hpp"
 #include "ngraph/runtime/backend.hpp"
 #include "ngraph/runtime/cpu/cpu_builder.hpp"
-#include "ngraph/runtime/tensor_view.hpp"
+#include "ngraph/runtime/tensor.hpp"
 
 using namespace std;
 using namespace ngraph;
