+++ src/ngraph/runtime/cpu/builder/reduce_function_window.cpp
@@ -18,7 +18,7 @@
 #include "ngraph/op/reduce_window.hpp"
 #include "ngraph/runtime/cpu/cpu_builder.hpp"
 #include "ngraph/runtime/cpu/cpu_external_function.hpp"
-#include "ngraph/runtime/tensor_view.hpp"
+#include "ngraph/runtime/tensor.hpp"
 
 using namespace std;
 using namespace ngraph;
