+++ src/ngraph/runtime/gpu/gpu_tensor_view.hpp
@@ -19,7 +19,7 @@
 #include <cuda.h>
 #include <memory>
 
-#include "ngraph/runtime/tensor_view.hpp"
+#include "ngraph/runtime/tensor.hpp"
 #include "ngraph/type/element_type.hpp"
 
 namespace ngraph
