+++ src/ngraph/runtime/host_tensor_view.hpp
@@ -18,7 +18,7 @@
 
 #include <memory>
 
-#include "ngraph/runtime/tensor_view.hpp"
+#include "ngraph/runtime/tensor.hpp"
 #include "ngraph/type/element_type.hpp"
 
 namespace ngraph
